SDK-4093: Added SecurityUpdateChecker / checks renaming
