+++ config/services.yaml
@@ -85,17 +85,17 @@ services:
                 - '@SprykerSdk\Evaluator\Checker\MultidimensionalArrayChecker\NestingStructure\ArrayMergeNestingStructure'
                 - '@SprykerSdk\Evaluator\Checker\MultidimensionalArrayChecker\NestingStructure\ReturnArrayNestingStructure'
 
-    SprykerSdk\Evaluator\Checker\SecurityChecker\SecurityChecker:
+    SprykerSdk\Evaluator\Checker\OpenSourceVulnerabilitiesChecker\OpenSourceVulnerabilitiesChecker:
         arguments:
             - '@symfony.console.application'
             - '@SprykerSdk\Evaluator\Resolver\PathResolver'
-            - '%security_checker_doc_url%'
+            - '%open_source_vulnerabilities_checker_doc_url%'
 
-    SprykerSdk\Evaluator\Checker\SecurityChecker\SecurityUpdateChecker:
+    SprykerSdk\Evaluator\Checker\SprykerSecurityChecker\SprykerSecurityChecker:
         arguments:
             - '@SprykerSdk\Evaluator\Reader\ComposerReaderInterface'
             - '@SprykerSdk\Evaluator\ReleaseApp\Infrastructure\Service\ReleaseAppService'
-            - '%security_update_checker_doc_url%'
+            - '%spryker_security_checker_doc_url%'
 
     SprykerSdk\Evaluator\Checker\DependencyProviderAdditionalLogicChecker\DependencyProviderAdditionalLogicChecker:
         arguments:
